move test_para_json to module base
