+++ source/module_io/test/CMakeLists.txt
@@ -142,14 +142,6 @@ AddTest(
   SOURCES csr_reader_test.cpp ../csr_reader.cpp ../file_reader.cpp ../sparse_matrix.cpp
 )
 
-if(ENABLE_RAPIDJSON)
-  AddTest(
-    TARGET io_para_json_test
-    LIBS ${math_libs} base device
-    SOURCES para_json_test.cpp ../input.cpp
-  )
-endif()
-
 AddTest(
   TARGET read_rhog_test
   LIBS ${math_libs} base device planewave
